Move current enrichers to /Enrichers
