+++ src/Arcus.Observability.Tests.Unit/Telemetry/KubernetesEnricherTests.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using Arcus.Observability.Telemetry.Serilog;
+using Arcus.Observability.Telemetry.Serilog.Enrichers;
 using Serilog;
 using Serilog.Events;
 using Xunit;
