+++ src/Arcus.Observability.Tests.Unit/Telemetry/VersionEnricherTests.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Linq;
 using Arcus.Observability.Telemetry.Serilog;
+using Arcus.Observability.Telemetry.Serilog.Enrichers;
 using Serilog;
 using Serilog.Core;
 using Serilog.Events;
